Moved methods into separate files
